rename to compiler.[c|h]
