+++ cyarg/compiler.h
@@ -1,5 +1,5 @@
-#ifndef cyarg_ast_compiler_h
-#define cyarg_ast_compiler_h
+#ifndef cyarg_compiler_h
+#define cyarg_compiler_h
 
 #include "object.h"
 #include "vm.h"
