+++ cyarg/parser_compiler.c
@@ -1,0 +1,981 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <math.h>
+
+#include "common.h"
+#include "parser.h"
+#include "compiler.h"
+#include "compiler_common.h"
+#include "memory.h"
+#include "scanner.h"
+
+typedef void (*ParseFn)(bool canAssign);
+
+typedef struct {
+    ParseFn prefix;
+    ParseFn infix;
+    Precedence precedence;
+} ParseRule;
+
+typedef struct {
+    Token name;
+    int depth;
+    bool isCaptured;
+} Local;
+
+typedef struct {
+    uint8_t index;
+    bool isLocal;
+} Upvalue;
+
+typedef struct Compiler {
+    struct Compiler* enclosing;
+    ObjFunction* function;
+    FunctionType type;
+
+    Local locals[UINT8_COUNT];
+    int localCount;
+    Upvalue upvalues[UINT8_COUNT];
+    int scopeDepth;
+} Compiler;
+
+typedef struct ClassCompiler {
+    struct ClassCompiler* enclosing;
+    bool hasSuperclass;
+} ClassCompiler;
+
+
+Compiler* current = NULL;
+ClassCompiler* currentClass = NULL;
+
+static Chunk* currentChunk() {
+    return &current->function->chunk;
+}
+
+
+static void emitByte(uint8_t byte) {
+    writeChunk(currentChunk(), byte, parser.previous.line);
+}
+
+static void emitBytes(uint8_t byte1, uint8_t byte2) {
+    emitByte(byte1);
+    emitByte(byte2);
+}
+
+static void emitLoop(int loopStart) {
+    emitByte(OP_LOOP);
+
+    int offset = currentChunk()->count - loopStart + 2;
+    if (offset > UINT16_MAX) error("Loop body too large.");
+
+    emitByte((offset >> 8) & 0xff);
+    emitByte(offset & 0xff);
+}
+
+static int emitJump(uint8_t instruction) {
+    emitByte(instruction);
+    emitByte(0xff);
+    emitByte(0xff);
+    return currentChunk()->count - 2;
+}
+
+static void emitReturn() {
+    if (current->type == TYPE_INITIALIZER) {
+        emitBytes(OP_GET_LOCAL, 0);
+    } else {
+        emitByte(OP_NIL);
+    }
+
+    emitByte(OP_RETURN);
+}
+
+static uint8_t makeConstant(Value value) {
+
+    int constant = addConstant(currentChunk(), value);
+    if (constant > UINT8_MAX) {
+        error("Too many constants in one chunk.");
+        return 0;
+    }
+
+    return (uint8_t)constant;
+}
+
+static int emitConstant(Value value) {
+
+    int32_t i = AS_INTEGER(value);
+    if (IS_INTEGER(value) && i <= INT8_MAX && i >= INT8_MIN) {
+        emitBytes(OP_IMMEDIATE, (uint8_t)i);
+    } else {
+        emitBytes(OP_CONSTANT, makeConstant(value));
+    }
+    return 2;
+}
+
+static void patchConstant(Value value, int offset) {
+
+    int32_t i = AS_INTEGER(value);
+    if (IS_INTEGER(value) && i <= INT8_MAX && i >= INT8_MIN) {
+        currentChunk()->code[offset] = OP_IMMEDIATE;
+        currentChunk()->code[offset + 1] = (uint8_t) i;
+    } else {
+        currentChunk()->code[offset] = OP_CONSTANT;
+        currentChunk()->code[offset + 1] = makeConstant(value);
+    }
+}
+
+static void patchJump(int offset) {
+    // -2 to adjust for the bytecode for the jump offset itself.
+    int jump = currentChunk()->count - offset - 2;
+
+    if (jump > UINT16_MAX) {
+        error("Too much code to jump over.");
+    }
+
+    currentChunk()->code[offset] = (jump >> 8) & 0xff;
+    currentChunk()->code[offset + 1] = jump & 0xff;
+}
+
+static void initCompiler(Compiler* compiler, FunctionType type) {
+    compiler->enclosing = current;
+    compiler->function = NULL;
+    compiler->type = type;
+    compiler->localCount = 0;
+    compiler->scopeDepth = 0;
+    compiler->function = newFunction();
+    current = compiler;
+    if (type != TYPE_SCRIPT) {
+        current->function->name = copyString(parser.previous.start,
+                                             parser.previous.length);
+    }
+
+    Local* local = &current->locals[current->localCount++];
+    local->depth = 0;
+    local->isCaptured = false;
+    if (type != TYPE_FUNCTION) {
+        local->name.start = "this";
+        local->name.length = 4;
+    } else {
+        local->name.start = "";
+        local->name.length = 0;
+    }
+}
+
+static ObjFunction* endCompiler() {
+    emitReturn();
+
+    ObjFunction* function = current->function;
+    current = current->enclosing;
+    return function;
+}
+
+static void beginScope() {
+    current->scopeDepth++;
+}
+
+static void endScope() {
+    current->scopeDepth--;
+
+    while (current->localCount > 0 &&
+           current->locals[current->localCount - 1].depth > current->scopeDepth) {
+        if (current->locals[current->localCount - 1].isCaptured) {
+            emitByte(OP_CLOSE_UPVALUE);
+        } else {
+            emitByte(OP_POP);
+        }
+        current->localCount--;
+    }
+}
+
+static void expression();
+static ParseRule* getRule(TokenType type);
+static void parsePrecedence(Precedence precedence);
+
+static uint8_t identifierConstant(Token* name) {
+    return makeConstant(OBJ_VAL(copyString(name->start, name->length)));
+}
+
+static bool identifiersEqual(Token* a, Token * b) {
+    if (a->length != b->length) return false;
+    return memcmp(a->start, b->start, a->length) == 0;
+}
+
+static int resolveLocal(Compiler* compiler, Token* name) {
+    for (int i = compiler->localCount - 1; i >= 0; i--) {
+        Local* local = &compiler->locals[i];
+        if (identifiersEqual(name, &local->name)) {
+            if (local->depth == -1) {
+                error("Can't read local variable in its own initializer.");
+            }
+            return i;
+        }
+    }
+
+    return -1;
+}
+
+static int addUpvalue(Compiler* compiler, uint8_t index, bool isLocal) {
+    int upvalueCount = compiler->function->upvalueCount;
+
+    for (int i = 0; i < upvalueCount; i++) {
+        Upvalue* upvalue = &compiler->upvalues[i];
+        if (upvalue->index == index && upvalue->isLocal == isLocal) {
+            return i;
+        }
+    }
+
+    if (upvalueCount == UINT8_COUNT) {
+        error("Too many closure variables in function.");
+        return 0;
+    }
+
+    compiler->upvalues[upvalueCount].isLocal = isLocal;
+    compiler->upvalues[upvalueCount].index = index;
+    return compiler->function->upvalueCount++;
+}
+
+static int resolveUpvalue(Compiler* compiler, Token* name) {
+    if (compiler->enclosing == NULL) return -1;
+
+    int local = resolveLocal(compiler->enclosing, name);
+    if (local != -1) {
+        compiler->enclosing->locals[local].isCaptured = true;
+        return addUpvalue(compiler, (uint8_t)local, true);
+    }
+
+    int upvalue = resolveUpvalue(compiler->enclosing, name);
+    if (upvalue != -1) {
+        return addUpvalue(compiler, (uint8_t)upvalue, false);
+    }
+
+    return -1;
+}
+
+static void addLocal(Token name) {
+    if (current->localCount == UINT8_COUNT) {
+        error("Too many local variables in function.");
+        return;
+    }
+
+    Local* local = &current->locals[current->localCount++];
+    local->name = name;
+    local->depth = -1;
+    local->isCaptured = false;
+}
+
+static void declareVariable() {
+    if (current->scopeDepth == 0) return;
+
+    Token* name = &parser.previous;
+    for (int i = current->localCount - 1; i >= 0; i--) {
+        Local* local = &current->locals[i];
+        if (local->depth != -1 && local->depth < current->scopeDepth) {
+            break;
+        }
+
+        if (identifiersEqual(name, &local->name)) {
+            error("Already a variable with this name in this scope.");
+        }
+    }
+
+    addLocal(*name);
+}
+
+static uint8_t parseVariable(const char* errorMessage) {
+    consume(TOKEN_IDENTIFIER, errorMessage);
+
+    declareVariable();
+    if (current->scopeDepth > 0) return 0;
+
+    return identifierConstant(&parser.previous);
+}
+
+static void markInitialized() {
+    if (current->scopeDepth == 0) return;
+    current->locals[current->localCount - 1].depth = current->scopeDepth;
+}
+
+static void defineVariable(uint8_t global) {
+    if (current->scopeDepth > 0) {
+        markInitialized();
+        return;
+    }
+
+    emitBytes(OP_DEFINE_GLOBAL, global);
+}
+
+static uint8_t argumentList() {
+    uint8_t argCount = 0;
+    if (!check(TOKEN_RIGHT_PAREN)) {
+        do {
+            expression();
+            if (argCount == 255) {
+                error("Can't have more than 255 arguments.");
+            }
+            argCount++;
+        } while (match(TOKEN_COMMA));
+    }
+    consume(TOKEN_RIGHT_PAREN, "Expect ')' aftger arguments.");
+    return argCount;
+}
+
+static uint8_t arrayInitExpressionsList() {
+    uint8_t expCount = 0;
+    if (!check(TOKEN_RIGHT_SQUARE_BRACKET)) {
+        do {
+            emitConstant(INTEGER_VAL(expCount));
+            expression();
+            if (expCount == 255) {
+                error("Can't have more than 255 array initialisers.");
+            }
+            emitByte(OP_SET_ELEMENT);
+            expCount++;
+        } while (match(TOKEN_COMMA));
+    }
+    return expCount;
+}
+
+static void and_(bool canAssign) {
+    int endJump = emitJump(OP_JUMP_IF_FALSE);
+
+    emitByte(OP_POP);
+    parsePrecedence(PREC_AND);
+
+    patchJump(endJump);
+}
+
+static void binary(bool canAssign) {
+    TokenType operatorType = parser.previous.type;
+    ParseRule* rule = getRule(operatorType);
+    parsePrecedence((Precedence)(rule->precedence + 1));
+    switch (operatorType) {
+        case TOKEN_BANG_EQUAL:    emitBytes(OP_EQUAL, OP_NOT); break;
+        case TOKEN_EQUAL_EQUAL:   emitByte(OP_EQUAL); break;
+        case TOKEN_GREATER:       emitByte(OP_GREATER); break;
+        case TOKEN_GREATER_EQUAL: emitBytes(OP_LESS, OP_NOT); break;
+        case TOKEN_RIGHT_SHIFT:   emitByte(OP_RIGHT_SHIFT); break;
+        case TOKEN_LESS:          emitByte(OP_LESS); break;
+        case TOKEN_LESS_EQUAL:    emitBytes(OP_GREATER, OP_NOT); break;
+        case TOKEN_LEFT_SHIFT:    emitByte(OP_LEFT_SHIFT); break;
+        case TOKEN_PLUS:          emitByte(OP_ADD); break;
+        case TOKEN_MINUS:         emitByte(OP_SUBTRACT); break;
+        case TOKEN_STAR:          emitByte(OP_MULTIPLY); break;
+        case TOKEN_SLASH:         emitByte(OP_DIVIDE); break;
+        case TOKEN_BAR:           emitByte(OP_BITOR); break;
+        case TOKEN_AMP:           emitByte(OP_BITAND); break;
+        case TOKEN_CARET:         emitByte(OP_BITXOR); break;
+        case TOKEN_PERCENT:       emitByte(OP_MODULO); break;
+        default:
+            return; // Unreachable.
+    }
+}
+
+static void deref(bool canAssign) {
+
+    parsePrecedence(PREC_ASSIGNMENT);   // prevents assignment within []
+    consume(TOKEN_RIGHT_SQUARE_BRACKET, "Expect ] after expression.");
+
+    if (canAssign && match(TOKEN_EQUAL)) {
+        expression();
+        emitByte(OP_SET_ELEMENT);
+    } else {
+        emitByte(OP_ELEMENT);
+    }
+}
+
+static void call(bool canAssign) {
+    uint8_t argCount = argumentList();
+    emitBytes(OP_CALL, argCount);
+}
+
+static void dot(bool canAssign) {
+    consume(TOKEN_IDENTIFIER, "Expect property name after '.'.");
+    uint8_t name = identifierConstant(&parser.previous);
+
+    if (canAssign && match(TOKEN_EQUAL)) {
+        expression();
+        emitBytes(OP_SET_PROPERTY, name);
+    } else if (match(TOKEN_LEFT_PAREN)) {
+        uint8_t argCount = argumentList();
+        emitBytes(OP_INVOKE, name);
+        emitByte(argCount);
+    } else {
+        emitBytes(OP_GET_PROPERTY, name);
+    }
+}
+
+static void literal(bool canAssign) {
+    switch (parser.previous.type) {
+        case TOKEN_RPEEK: emitBytes(OP_GET_BUILTIN, BUILTIN_RPEEK); break;
+        case TOKEN_RPOKE: emitBytes(OP_GET_BUILTIN, BUILTIN_RPOKE); break;
+        case TOKEN_IMPORT: emitBytes(OP_GET_BUILTIN, BUILTIN_IMPORT); break;
+        case TOKEN_MAKE_ARRAY: emitBytes(OP_GET_BUILTIN, BUILTIN_MAKE_ARRAY); break;
+        case TOKEN_MAKE_ROUTINE: emitBytes(OP_GET_BUILTIN, BUILTIN_MAKE_ROUTINE); break;
+        case TOKEN_MAKE_CHANNEL: emitBytes(OP_GET_BUILTIN, BUILTIN_MAKE_CHANNEL); break;
+        case TOKEN_RESUME: emitBytes(OP_GET_BUILTIN, BUILTIN_RESUME); break;
+        case TOKEN_SEND: emitBytes(OP_GET_BUILTIN, BUILTIN_SEND); break;
+        case TOKEN_RECEIVE: emitBytes(OP_GET_BUILTIN, BUILTIN_RECEIVE); break;        
+        case TOKEN_SHARE: emitBytes(OP_GET_BUILTIN, BUILTIN_SHARE); break;
+        case TOKEN_START: emitBytes(OP_GET_BUILTIN, BUILTIN_START); break;
+        case TOKEN_PEEK: emitBytes(OP_GET_BUILTIN, BUILTIN_PEEK); break;
+        case TOKEN_LEN: emitBytes(OP_GET_BUILTIN, BUILTIN_LEN); break;
+        case TOKEN_FALSE: emitByte(OP_FALSE); break;
+        case TOKEN_NIL: emitByte(OP_NIL); break;
+        case TOKEN_TRUE: emitByte(OP_TRUE); break;
+        case TOKEN_MACHINE_UINT32: emitByte(OP_TYPE_LITERAL); break;
+        default: return; // Unreachable.
+    }
+}
+
+static void grouping(bool canAssign) {
+    expression();
+    consume(TOKEN_RIGHT_PAREN, "Expect ')' after expression.");
+}
+
+static void arrayinit(bool canAssign) {
+
+    emitBytes(OP_GET_BUILTIN, BUILTIN_MAKE_ARRAY);
+    
+    int constantSize = emitConstant(INTEGER_VAL(0));
+    int constantOffset = currentChunk()->count - constantSize;
+
+    emitBytes(OP_CALL, 1);
+
+    uint8_t items = arrayInitExpressionsList();
+    consume(TOKEN_RIGHT_SQUARE_BRACKET, "Expect ']' initialising array.");
+
+    patchConstant(INTEGER_VAL(items), constantOffset);
+}
+
+static uint32_t strtoNum(const char* literal, int length, int radix) {
+    uint32_t val = 0;
+    for (int i = length - 1 ; i >= 0; i--) {
+        uint32_t positionVal = 0;
+        switch (literal[i]) {
+            case '0': positionVal = 0; break;
+            case '1': positionVal = 1; break;
+            case '2': positionVal = 2; break;
+            case '3': positionVal = 3; break;
+            case '4': positionVal = 4; break;
+            case '5': positionVal = 5; break;
+            case '6': positionVal = 6; break;
+            case '7': positionVal = 7; break;
+            case '8': positionVal = 8; break;
+            case '9': positionVal = 9; break;
+            case 'A': positionVal = 10; break;
+            case 'B': positionVal = 11; break;
+            case 'C': positionVal = 12; break;
+            case 'D': positionVal = 13; break;
+            case 'E': positionVal = 14; break;
+            case 'F': positionVal = 15; break;
+            case 'a': positionVal = 10; break;
+            case 'b': positionVal = 11; break;
+            case 'c': positionVal = 12; break;
+            case 'd': positionVal = 13; break;
+            case 'e': positionVal = 14; break;
+            case 'f': positionVal = 15; break;
+        }
+        uint32_t power = length - 1 - i;
+        val += positionVal * pow(radix, power);
+    }
+    return val;
+}
+
+static void number(bool canAssign) {
+    int radix = 0;
+    bool sign_bit = true;
+    const char* number_start = parser.previous.start;
+    int number_len = parser.previous.length;
+    if (parser.previous.length > 1) {
+        switch (parser.previous.start[1]) {
+            case 'x': radix = 16; sign_bit = false; break;
+            case 'X': radix = 16; sign_bit = false; break;
+            case 'b': radix = 2; sign_bit = false; break;
+            case 'B': radix = 2; sign_bit = false; break;
+            case 'd': radix = 10; sign_bit = false; break;
+            case 'D': radix = 10; sign_bit = false; break;
+        }
+        if (radix != 0) {
+            number_start = &parser.previous.start[2];
+            number_len -= 2;
+        }
+    }
+    if (radix == 0) {
+        radix = 10;
+    }
+
+    if (radix == 10 && sign_bit) {
+        if (memchr(number_start, '.', number_len)) {
+            // for now, use C's stdlib to reuse double formatting.
+            double value = strtod(number_start, NULL);
+            emitConstant(DOUBLE_VAL(value));
+        } else {
+            int value = atoi(number_start);
+            emitConstant(INTEGER_VAL(value));
+        }
+    }
+    else {
+        uint32_t value = strtoNum(number_start, number_len, radix);
+        emitConstant(UINTEGER_VAL(value));
+    }
+}
+
+static void or_(bool canAssign) {
+    int elseJump = emitJump(OP_JUMP_IF_FALSE);
+    int endJump = emitJump(OP_JUMP);
+
+    patchJump(elseJump);
+    emitByte(OP_POP);
+
+    parsePrecedence(PREC_OR);
+    patchJump(endJump);
+}
+
+static void string(bool canAssign) {
+    emitConstant(OBJ_VAL(copyString(parser.previous.start + 1,
+                                    parser.previous.length - 2)));
+}
+
+static void namedVariable(Token name, bool canAssign) {
+    uint8_t getOp, setOp;
+    int arg = resolveLocal(current, &name);
+    if (arg != -1) {
+        getOp = OP_GET_LOCAL;
+        setOp = OP_SET_LOCAL;
+    } else if ((arg = resolveUpvalue(current, &name)) != -1) {
+        getOp = OP_GET_UPVALUE;
+        setOp = OP_SET_UPVALUE;
+    } else {
+        arg = identifierConstant(&name);
+        getOp = OP_GET_GLOBAL;
+        setOp = OP_SET_GLOBAL;
+    }
+    
+    if (canAssign && match(TOKEN_EQUAL)) {
+        expression();
+        emitBytes(setOp, (uint8_t)arg);
+    } else {
+        emitBytes(getOp, (uint8_t)arg);
+    }
+}
+
+static void variable(bool canAssign) {
+    namedVariable(parser.previous, canAssign);
+}
+
+static Token syntheticToken(const char* text) {
+    Token token;
+    token.start = text;
+    token.length = (int)strlen(text);
+    return token;
+}
+
+static void super_(bool canAssign) {
+    if (currentClass == NULL) {
+        error("Can't use 'super' outside of a class.");
+    } else if (!currentClass->hasSuperclass) {
+        error("Can't use 'super' in a class with no superclass.");
+    }
+
+    consume(TOKEN_DOT, "Expect '.' after 'super'.");
+    consume(TOKEN_IDENTIFIER, "Expect superclass method name.");
+    uint8_t name = identifierConstant(&parser.previous);
+
+    namedVariable(syntheticToken("this"), false);
+    if (match(TOKEN_LEFT_PAREN)) {
+        uint8_t argCount = argumentList();
+        namedVariable(syntheticToken("super"), false);
+        emitBytes(OP_SUPER_INVOKE, name);
+        emitByte(argCount);
+    } else {
+        namedVariable(syntheticToken("super"), false);
+        emitBytes(OP_GET_SUPER, name);
+    }
+}
+
+static void this_(bool canAssign) {
+    if (currentClass == NULL) {
+        error("Can't use 'this' outside of a class.");
+        return;
+    }
+
+    variable(false);
+}
+
+static void unary(bool canAssign) {
+    TokenType operatorType = parser.previous.type;
+
+    // Compile the operand.
+    parsePrecedence(PREC_UNARY);
+
+    // Emit the operator instruction.
+    switch (operatorType) {
+        case TOKEN_BANG: emitByte(OP_NOT); break;
+        case TOKEN_MINUS: emitByte(OP_NEGATE); break;
+        default: return; // Unreachable.
+    }
+}
+
+static ParseRule rules[] = {
+    [TOKEN_LEFT_PAREN]           = {grouping,  call,   PREC_CALL},
+    [TOKEN_RIGHT_PAREN]          = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_LEFT_BRACE]           = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_RIGHT_BRACE]          = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_LEFT_SQUARE_BRACKET]  = {arrayinit, deref,  PREC_DEREF},
+    [TOKEN_RIGHT_SQUARE_BRACKET] = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_COMMA]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_DOT]                  = {NULL,      dot,    PREC_CALL},
+    [TOKEN_MINUS]                = {unary,     binary, PREC_TERM},
+    [TOKEN_PLUS]                 = {NULL,      binary, PREC_TERM},
+    [TOKEN_SEMICOLON]            = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_SLASH]                = {NULL,      binary, PREC_FACTOR},
+    [TOKEN_STAR]                 = {NULL,      binary, PREC_FACTOR},
+    [TOKEN_BAR]                  = {NULL,      binary, PREC_TERM},
+    [TOKEN_AMP]                  = {NULL,      binary, PREC_TERM},
+    [TOKEN_PERCENT]              = {NULL,      binary, PREC_TERM},
+    [TOKEN_CARET]                = {NULL,      binary, PREC_TERM},
+    [TOKEN_BANG]                 = {unary,     NULL,   PREC_NONE},
+    [TOKEN_BANG_EQUAL]           = {NULL,      binary, PREC_EQUALITY},
+    [TOKEN_EQUAL]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_EQUAL_EQUAL]          = {NULL,      binary, PREC_EQUALITY},
+    [TOKEN_GREATER]              = {NULL,      binary, PREC_COMPARISON},
+    [TOKEN_GREATER_EQUAL]        = {NULL,      binary, PREC_COMPARISON},
+    [TOKEN_RIGHT_SHIFT]          = {NULL,      binary, PREC_FACTOR},
+    [TOKEN_LESS]                 = {NULL,      binary, PREC_COMPARISON},
+    [TOKEN_LESS_EQUAL]           = {NULL,      binary, PREC_COMPARISON},
+    [TOKEN_LEFT_SHIFT]           = {NULL,      binary, PREC_FACTOR},
+    [TOKEN_IDENTIFIER]           = {variable,  NULL,   PREC_NONE},
+    [TOKEN_STRING]               = {string,    NULL,   PREC_NONE},
+    [TOKEN_NUMBER]               = {number,    NULL,   PREC_NONE},
+    [TOKEN_AND]                  = {NULL,      and_,   PREC_AND},
+    [TOKEN_CLASS]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_ELSE]                 = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_FALSE]                = {literal,   NULL,   PREC_NONE},
+    [TOKEN_FOR]                  = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_FUN]                  = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_IF]                   = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_IMPORT]               = {literal,   NULL,   PREC_NONE},
+    [TOKEN_LEN]                  = {literal,   NULL,   PREC_NONE},
+    [TOKEN_MACHINE_UINT32]       = {literal,   NULL,   PREC_NONE},
+    [TOKEN_MAKE_ARRAY]           = {literal,   NULL,   PREC_NONE},
+    [TOKEN_MAKE_CHANNEL]         = {literal,   NULL,   PREC_NONE},
+    [TOKEN_MAKE_ROUTINE]         = {literal,   NULL,   PREC_NONE},
+    [TOKEN_NIL]                  = {literal,   NULL,   PREC_NONE},
+    [TOKEN_OR]                   = {NULL,      or_,    PREC_OR},
+    [TOKEN_PEEK]                 = {literal,   NULL,   PREC_NONE},
+    [TOKEN_PRINT]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_RECEIVE]              = {literal,   NULL,   PREC_NONE},
+    [TOKEN_RESUME]               = {literal,   NULL,   PREC_NONE},
+    [TOKEN_RETURN]               = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_RPEEK]                = {literal,   NULL,   PREC_NONE},
+    [TOKEN_RPOKE]                = {literal,   NULL,   PREC_NONE},
+    [TOKEN_SEND]                 = {literal,   NULL,   PREC_NONE},
+    [TOKEN_SHARE]                = {literal,   NULL,   PREC_NONE},
+    [TOKEN_START]                = {literal,   NULL,   PREC_NONE},
+    [TOKEN_SET_ELEMENT]          = {literal,   NULL,   PREC_NONE},
+    [TOKEN_SUPER]                = {super_,    NULL,   PREC_NONE},
+    [TOKEN_THIS]                 = {this_,     NULL,   PREC_NONE},
+    [TOKEN_TRUE]                 = {literal,   NULL,   PREC_NONE},
+    [TOKEN_VAR]                  = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_WHILE]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_YIELD]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_ERROR]                = {NULL,      NULL,   PREC_NONE},
+    [TOKEN_EOF]                  = {NULL,      NULL,   PREC_NONE},
+};
+
+static void parsePrecedence(Precedence precedence) {
+    advance();
+    ParseFn prefixRule = getRule(parser.previous.type)->prefix;
+    if (prefixRule == NULL) {
+        error("Expect expression.");
+        return;
+    }
+
+    bool canAssign = precedence <= PREC_ASSIGNMENT;
+    prefixRule(canAssign);
+
+    while (precedence <= getRule(parser.current.type)->precedence) {
+        advance();
+        ParseFn infixRule = getRule(parser.previous.type)->infix;
+        infixRule(canAssign);
+    }
+
+    if (canAssign && match(TOKEN_EQUAL)) {
+        error("Invalid assignment target.");
+    }
+}
+
+static ParseRule* getRule(TokenType type) {
+    return &rules[type];
+}
+
+static void expression() {
+    parsePrecedence(PREC_ASSIGNMENT);
+}
+
+static void block() {
+    while (!check(TOKEN_RIGHT_BRACE) && !check(TOKEN_EOF)) {
+        declaration();
+    }
+
+    consume(TOKEN_RIGHT_BRACE, "Expect '}' after block.");
+}
+
+static void function(FunctionType type) {
+    Compiler compiler;
+    initCompiler(&compiler, type);
+    beginScope();
+
+    consume(TOKEN_LEFT_PAREN, "Expect '(' after function name.");
+    if (!check(TOKEN_RIGHT_PAREN)) {
+        do {
+            current->function->arity++;
+            if (current->function->arity > 255) {
+                errorAtCurrent("Can't have more than 255 parameters.");
+            }
+            uint8_t constant = parseVariable("Expect parameter name.");
+            defineVariable(constant);
+        } while (match(TOKEN_COMMA));
+    }
+    consume(TOKEN_RIGHT_PAREN, "Expect ')' after parameters.");
+    consume(TOKEN_LEFT_BRACE, "Expect '{' before function body.");
+    block();
+
+    ObjFunction* function = endCompiler();
+    emitBytes(OP_CLOSURE, makeConstant(OBJ_VAL(function)));
+
+    for (int i = 0; i < function->upvalueCount; i++) {
+        emitByte(compiler.upvalues[i].isLocal ? 1 : 0);
+        emitByte(compiler.upvalues[i].index);
+    }
+}
+
+static void method() {
+    consume(TOKEN_IDENTIFIER, "Expect method name.");
+    uint8_t constant = identifierConstant(&parser.previous);
+
+    FunctionType type = TYPE_METHOD;
+    if (parser.previous.length == 4 &&
+        memcmp(parser.previous.start, "init", 4) == 0) {
+        type = TYPE_INITIALIZER;
+    }
+    
+    function(type);
+
+    emitBytes(OP_METHOD, constant);
+}
+
+static void classDeclaration() {
+    consume(TOKEN_IDENTIFIER, "Expect class name.");
+    Token className = parser.previous;
+    uint8_t nameConstant = identifierConstant(&parser.previous);
+    declareVariable();
+
+    emitBytes(OP_CLASS, nameConstant);
+    defineVariable(nameConstant);
+
+    ClassCompiler classCompiler;
+    classCompiler.hasSuperclass = false;
+    classCompiler.enclosing = currentClass;
+    currentClass = &classCompiler;
+
+    if (match(TOKEN_LESS)) {
+        consume(TOKEN_IDENTIFIER, "Expect superclass name.");
+        variable(false);
+
+        if (identifiersEqual(&className, &parser.previous)) {
+            error("A class can't inherit from itself.");
+        }
+
+        beginScope();
+        addLocal(syntheticToken("super"));
+        defineVariable(0);
+
+        namedVariable(className, false);
+        emitByte(OP_INHERIT);
+        classCompiler.hasSuperclass = true;
+    }
+
+    // Make the class name conveniently available on the stack for method definition.
+    namedVariable(className, false);
+    consume(TOKEN_LEFT_BRACE, "Expect '{' before class body.");
+    while (!check(TOKEN_RIGHT_BRACE) && !check(TOKEN_EOF)) {
+        method();
+    }
+    consume(TOKEN_RIGHT_BRACE, "Expect '}' after class body.");
+    emitByte(OP_POP);
+
+    if (classCompiler.hasSuperclass) {
+        endScope();
+    }
+
+    currentClass = currentClass->enclosing;
+}
+
+static void funDeclaration() {
+    uint8_t global = parseVariable("Expect function name.");
+    markInitialized();
+    function(TYPE_FUNCTION);
+    defineVariable(global);
+}
+
+static void varDeclaration() {
+    uint8_t global = parseVariable("Expect variable name.");
+
+    if (match(TOKEN_EQUAL)) {
+        expression();
+    } else {
+        emitByte(OP_NIL);
+    }
+    consume(TOKEN_SEMICOLON, "Expect ';' after variable declaration");
+
+    defineVariable(global);
+}
+
+static void expressionStatement() {
+    expression();
+    consume(TOKEN_SEMICOLON, "Expect ';' after expression.");
+    emitByte(OP_POP);
+}
+
+static void forStatement() {
+    beginScope();
+    consume(TOKEN_LEFT_PAREN, "Expect '(' after 'for'.");
+    if (match(TOKEN_SEMICOLON)) {
+        // No initializer.
+    } else if (match(TOKEN_VAR)) {
+        varDeclaration();
+    } else {
+        expressionStatement();
+    }
+
+    int loopStart = currentChunk()->count;
+    int exitJump = -1;
+    if (!match(TOKEN_SEMICOLON)) {
+        expression();
+        consume(TOKEN_SEMICOLON, "Expect ';' after loop condition.");
+
+        // Jump out of the loop if the condition is false.
+        exitJump = emitJump(OP_JUMP_IF_FALSE);
+        emitByte(OP_POP); // Condition.
+    }
+
+    if (!match(TOKEN_RIGHT_PAREN)) {
+        int bodyJump = emitJump(OP_JUMP);
+        int incrementStart = currentChunk()->count;
+        expression();
+        emitByte(OP_POP);
+        consume(TOKEN_RIGHT_PAREN, "Expect ')' after for clauses.");
+
+        emitLoop(loopStart);
+        loopStart = incrementStart;
+        patchJump(bodyJump);
+    }
+
+    statement();
+    emitLoop(loopStart);
+
+    if (exitJump != -1) {
+        patchJump(exitJump);
+        emitByte(OP_POP);
+    }
+
+    endScope();
+}
+
+static void ifStatement() {
+    consume(TOKEN_LEFT_PAREN, "Expect '(' after 'if'.");
+    expression();
+    consume(TOKEN_RIGHT_PAREN, "Expect ')' after condition.");
+
+    int thenJump = emitJump(OP_JUMP_IF_FALSE);
+    emitByte(OP_POP);
+    statement();
+
+    int elseJump = emitJump(OP_JUMP);
+
+    patchJump(thenJump);
+    emitByte(OP_POP);
+
+    if (match(TOKEN_ELSE)) statement();
+    patchJump(elseJump);
+}
+
+static void printStatement() {
+    expression();
+    consume(TOKEN_SEMICOLON, "Expect ';' after value.");
+    emitByte(OP_PRINT);
+}
+
+static void returnStatement() {
+    if (current->type == TYPE_SCRIPT) {
+        error("Can't return from top-level code.");
+    }
+    
+    if (match(TOKEN_SEMICOLON)) {
+        emitReturn();
+    } else {
+        if (current->type == TYPE_INITIALIZER) {
+            error("Can't return a value from an initializer.");
+        }
+
+        expression();
+        consume(TOKEN_SEMICOLON, "Expect ';' after return value.");
+        emitByte(OP_RETURN);
+    }
+}
+
+static void yieldStatement() {
+    if (current->type == TYPE_SCRIPT) {
+        error("Can't yield from top-level code.");
+    }
+    
+    if (match(TOKEN_SEMICOLON)) {
+        emitBytes(OP_NIL, OP_YIELD);
+    } else {
+
+        expression();
+        consume(TOKEN_SEMICOLON, "Expect ';' after yield value.");
+        emitByte(OP_YIELD);
+    }
+}
+
+static void whileStatement() {
+    int loopStart = currentChunk()->count;
+    consume(TOKEN_LEFT_PAREN, "Expect '(' after 'while'.");
+    expression();
+    consume(TOKEN_RIGHT_PAREN, "Expect ')' after condition.");
+
+    int exitJump = emitJump(OP_JUMP_IF_FALSE);
+    emitByte(OP_POP);
+    statement();
+    emitLoop(loopStart);
+
+    patchJump(exitJump);
+    emitByte(OP_POP);
+}
+
+
+ObjFunction* compile(const char* source) {
+    initScanner(source);
+    Compiler compiler;
+    initCompiler(&compiler, TYPE_SCRIPT);
+
+    parser.hadError = false;
+    parser.panicMode = false;
+
+    advance();
+    while (!match(TOKEN_EOF)) {
+        declaration();
+    }
+    ObjFunction* function = endCompiler();
+    return parser.hadError ? NULL : function;
+}
+
+void markCompilerRoots() {
+    Compiler* compiler = current;
+    while (compiler != NULL) {
+        markObject((Obj*)compiler->function);
+        compiler = compiler->enclosing;
+    }
+}
